addteams: script to add teams to a course
we still need to create repos also. and there are some issues
when rerunning the script on an existing teams/ directory when.
we need some way to avoid adding the same user to multiple teams
and we need to allow rerunning when some group have decided to
edit their group membership (adding/removing members)
(maybe removing should be manual.)