addteams: support creating repos, delete repos/teams, list repos.
We now support creating a repository with the same name as the
team who can contribute to the repo. (I think the owner is still
the organization, i.e. uis-dat320-fall2014 presently.)

Added support for listing repos (-list argument)

Added support for deleting repos and the corresponding team
(-delete argument)

Also fixed a silly bug in the way that the next group number was
calculated. Now, all group numbers should be sequential.