buildall script for initial scanning lab3 handins.
the script diffs the various parts of lab3 with the template
source code, to check if work has been done in the relevant
files. then we try to build the Go files to see if they compile.
the script will also check if lab3 has been downloaded; some
students have not.

should be run in the folder where teachers_pet close_repos was
performed.